+++ plugin/giveawayplugin/giveaway_test.go
@@ -1,0 +1,98 @@
+package giveawayplugin
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/torlenor/abylebotter/plugin"
+)
+
+func TestCreateGiveawayPlugin(t *testing.T) {
+	assert := assert.New(t)
+
+	p, err := New()
+	assert.NoError(err)
+	assert.Equal(nil, p.API)
+
+	api := plugin.MockAPI{}
+	p.SetAPI(&api)
+}
+
+func TestGiveawayPlugin_OnPost(t *testing.T) {
+	// assert := assert.New(t)
+
+	// p := GiveawayPlugin{
+	// 	randomizer: mockRandomizer{},
+	// }
+	// assert.Equal(nil, p.API)
+
+	// api := plugin.MockAPI{}
+	// p.SetAPI(&api)
+
+	// postToPlugin := model.Post{
+	// 	ChannelID: "CHANNEL ID",
+	// 	Channel:   "SOME CHANNEL",
+	// 	UserID:    "SOME USER ID",
+	// 	User:      "USER 1",
+	// 	Content:   "MESSAGE CONTENT",
+	// 	IsPrivate: false,
+	// }
+	// p.OnPost(postToPlugin)
+	// assert.Equal(false, api.WasCreatePostCalled)
+
+	// api.Reset()
+	// postToPlugin.Content = "!roll"
+	// expectedPostFromPlugin := model.Post{
+	// 	ChannelID: "CHANNEL ID",
+	// 	Channel:   "SOME CHANNEL",
+	// 	UserID:    "SOME USER ID",
+	// 	User:      "USER 1",
+	// 	Content:   "<@" + postToPlugin.UserID + "> rolled *" + strconv.Itoa(123) + "* in [0,100]",
+	// 	IsPrivate: false,
+	// }
+	// p.OnPost(postToPlugin)
+	// assert.Equal(true, api.WasCreatePostCalled)
+	// assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	// api.Reset()
+	// postToPlugin.Content = "!roll 1000"
+	// expectedPostFromPlugin = model.Post{
+	// 	ChannelID: "CHANNEL ID",
+	// 	Channel:   "SOME CHANNEL",
+	// 	UserID:    "SOME USER ID",
+	// 	User:      "USER 1",
+	// 	Content:   "<@" + postToPlugin.UserID + "> rolled *" + strconv.Itoa(123) + "* in [0,1000]",
+	// 	IsPrivate: false,
+	// }
+	// p.OnPost(postToPlugin)
+	// assert.Equal(true, api.WasCreatePostCalled)
+	// assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	// api.Reset()
+	// postToPlugin.Content = "!roll -1"
+	// expectedPostFromPlugin = model.Post{
+	// 	ChannelID: "CHANNEL ID",
+	// 	Channel:   "SOME CHANNEL",
+	// 	UserID:    "SOME USER ID",
+	// 	User:      "USER 1",
+	// 	Content:   "Number must be > 0",
+	// 	IsPrivate: false,
+	// }
+	// p.OnPost(postToPlugin)
+	// assert.Equal(true, api.WasCreatePostCalled)
+	// assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	// api.Reset()
+	// postToPlugin.Content = "!roll sdsadsad"
+	// expectedPostFromPlugin = model.Post{
+	// 	ChannelID: "CHANNEL ID",
+	// 	Channel:   "SOME CHANNEL",
+	// 	UserID:    "SOME USER ID",
+	// 	User:      "USER 1",
+	// 	Content:   "Not a number",
+	// 	IsPrivate: false,
+	// }
+	// p.OnPost(postToPlugin)
+	// assert.Equal(true, api.WasCreatePostCalled)
+	// assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+}