+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 **New plugins:**
 
 - Version Plugin: To the command *!version* the plugin will answer with the version of the Bot.
+- Giveaway Plugin: Lets you hold giveaways in your channel and let the bot pick a winner.
 
 ## [0.0.3](https://git.abyle.org/hps/abylebotter/-/tree/0.0.3) (2020-03-31)